+++ lib/cli/src/frameworks/vue/0-Welcome.stories.js
