+++ lib/cli/src/frameworks/vue/1-Button.stories.js
@@ -1,7 +1,7 @@
 import { action } from '@storybook/addon-actions';
 import { linkTo } from '@storybook/addon-links';
 
-import MyButton from './MyButton.vue';
+import MyButton from './Button.vue';
 
 export default {
   title: 'Button',
