+++ lib/cli/src/frameworks/vue/Welcome.vue
