+++ lib/cli/src/generators/SFC_VUE/index.ts
@@ -1,42 +1,10 @@
-import {
-  retrievePackageJson,
-  getVersionedPackages,
-  writePackageJson,
-  getBabelDependencies,
-  installDependencies,
-  copyTemplate,
-} from '../../helpers';
+import baseGenerator, { Generator } from '../generator';
 import { StoryFormat } from '../../project_types';
-import { Generator } from '../generator';
 
-const generator: Generator = async (npmOptions, { storyFormat }) => {
-  const packages = [
-    '@storybook/vue',
-    '@storybook/addon-actions',
-    '@storybook/addon-links',
-    '@storybook/addons',
-  ];
-  if (storyFormat === StoryFormat.MDX) {
-    packages.push('@storybook/addon-docs');
-  }
-  const versionedPackages = await getVersionedPackages(npmOptions, ...packages);
-
-  copyTemplate(__dirname, storyFormat);
-
-  const packageJson = await retrievePackageJson();
-
-  packageJson.dependencies = packageJson.dependencies || {};
-  packageJson.devDependencies = packageJson.devDependencies || {};
-
-  packageJson.scripts = packageJson.scripts || {};
-  packageJson.scripts.storybook = 'start-storybook -p 6006';
-  packageJson.scripts['build-storybook'] = 'build-storybook';
-
-  writePackageJson(packageJson);
-
-  const babelDependencies = await getBabelDependencies(npmOptions, packageJson);
-
-  installDependencies({ ...npmOptions, packageJson }, [...versionedPackages, ...babelDependencies]);
+const generator: Generator = async (npmOptions, options) => {
+  baseGenerator(npmOptions, options, 'vue', {
+    dirname: options.storyFormat === StoryFormat.MDX ? __dirname : undefined,
+  });
 };
 
 export default generator;
