--- lib/cli/src/generators/SFC_VUE/template-csf/.storybook/main.js
@@ -1,4 +1,0 @@
-module.exports = {
-  stories: ['../src/**/*.stories.js'],
-  addons: ['@storybook/addon-actions', '@storybook/addon-links'],
-};
