--- lib/cli/src/generators/SFC_VUE/template-mdx/.storybook/main.js
@@ -1,4 +1,0 @@
-module.exports = {
-  stories: ['../src/**/*.stories.(js|mdx)'],
-  addons: ['@storybook/addon-docs', '@storybook/addon-actions', '@storybook/addon-links'],
-};
