+++ lib/cli/src/generators/SFC_VUE/template-mdx/stories/0-Welcome.stories.mdx
