+++ lib/cli/src/generators/SFC_VUE/template-mdx/stories/1-Button.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story } from '@storybook/addon-docs/blocks';
 import { action } from '@storybook/addon-actions';
 import { linkTo } from '@storybook/addon-links';
-import MyButton from './MyButton.vue';
+import MyButton from './Button.vue';
 
 <Meta title="Button" component={MyButton} />
